server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bencoder>=0.2.0",
    "cloudscraper>=1.2.71",
    "flask>=3.1.2",
    "werkzeug==3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "mysql-connector-python>=9.4.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pymediainfo>=7.0.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "qbittorrent-api>=2025.7.0",
    "transmission-rpc>=7.0.11",
    "dotenv>=0.9.9",
    "pysocks>=1.7.1",
]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
